pyproject.toml
[tool.poetry]
name = "gcp-api-pub-sub-function-bigquery"
version = "0.1.0"
description = ""
authors = ["paribe <paribe@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
google-cloud-pubsub = "^2.27.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
